implemented facade as event and move facade class into extension facade class
