+++ src/Runner/Extension/Facade.php
@@ -10,84 +10,42 @@
 namespace PHPUnit\Runner\Extension;
 
 use PHPUnit\Event\EventFacadeIsSealedException;
-use PHPUnit\Event\Facade as EventFacade;
 use PHPUnit\Event\Subscriber;
 use PHPUnit\Event\Tracer\Tracer;
 use PHPUnit\Event\UnknownSubscriberTypeException;
 
-/**
- * @no-named-arguments Parameter names are not covered by the backward compatibility promise for PHPUnit
- */
-final class Facade
+interface Facade
 {
-    private bool $replacesOutput                 = false;
-    private bool $replacesProgressOutput         = false;
-    private bool $replacesResultOutput           = false;
-    private bool $requiresCodeCoverageCollection = false;
+    /**
+     * @throws EventFacadeIsSealedException
+     * @throws UnknownSubscriberTypeException
+     */
+    public function registerSubscribers(Subscriber ...$subscribers): void;
 
     /**
      * @throws EventFacadeIsSealedException
      * @throws UnknownSubscriberTypeException
      */
-    public function registerSubscribers(Subscriber ...$subscribers): void
-    {
-        EventFacade::instance()->registerSubscribers(...$subscribers);
-    }
-
-    /**
-     * @throws EventFacadeIsSealedException
-     * @throws UnknownSubscriberTypeException
-     */
-    public function registerSubscriber(Subscriber $subscriber): void
-    {
-        EventFacade::instance()->registerSubscriber($subscriber);
-    }
+    public function registerSubscriber(Subscriber $subscriber): void;
 
     /**
      * @throws EventFacadeIsSealedException
      */
-    public function registerTracer(Tracer $tracer): void
-    {
-        EventFacade::instance()->registerTracer($tracer);
-    }
+    public function registerTracer(Tracer $tracer): void;
 
-    public function replaceOutput(): void
-    {
-        $this->replacesOutput = true;
-    }
+    public function replaceOutput(): void;
 
-    public function replacesOutput(): bool
-    {
-        return $this->replacesOutput;
-    }
+    public function replacesOutput(): bool;
 
-    public function replaceProgressOutput(): void
-    {
-        $this->replacesProgressOutput = true;
-    }
+    public function replaceProgressOutput(): void;
 
-    public function replacesProgressOutput(): bool
-    {
-        return $this->replacesOutput || $this->replacesProgressOutput;
-    }
+    public function replacesProgressOutput(): bool;
 
-    public function replaceResultOutput(): void
-    {
-        $this->replacesResultOutput = true;
-    }
+    public function replaceResultOutput(): void;
 
-    public function replacesResultOutput(): bool
-    {
-        return $this->replacesOutput || $this->replacesResultOutput;
-    }
+    public function replacesResultOutput(): bool;
 
-    public function requireCodeCoverageCollection(): void
-    {
-        $this->requiresCodeCoverageCollection = true;
-    }
+    public function requireCodeCoverageCollection(): void;
 
-    public function requiresCodeCoverageCollection(): bool
-    {
-        return $this->requiresCodeCoverageCollection;
-    }
+    public function requiresCodeCoverageCollection(): bool;
 }
